Add files via upload
The three codes written here can be used for any optimization problem by changing the objective function and the constraints at the end of the code. The current objective function is written to design a minimum-volume plate girder for a highway bridge, which satisfies the requirements of the American Association of State Highway and Transportation Officials (AASHTO) Specifications. The dead load for the girder consists of the weight of the pavement and the girder’s self-weight. The live load consists of the equivalent uniform load and concentrated loads based on the HS-20(MS18) truck-loading condition.